pyproject.toml
[project]
name = "ai-sub-agents"
version = "0.1.0"
description = "Multi-agent job application assistant with live trace dashboard"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "pydantic>=2.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]
dashboard = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.0",
]

[project.scripts]
agent-war-room = "main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
